Fix moving tensors to correct device.
