+++ neural_editor/seq2seq/batch_utils.py
@@ -6,7 +6,7 @@ from torch.nn.utils.rnn import pad_sequence
 
 
 def split_sequences_on_hunks(sequences: Tensor, new_hunk_index: int,
-                             pad_index: int, sos_index: int, eos_index: int) -> Tuple[Tensor, Tensor]:
+                             pad_index: int, sos_index: int, eos_index: int, device) -> Tuple[Tensor, Tensor]:
     # put all examples in a row
     sequences_in_row = sequences.reshape(-1)
     # remove batch padding, SOS and EOS tokens
@@ -21,7 +21,7 @@ def split_sequences_on_hunks(sequences: Tensor, new_hunk_index: int,
     # split into hunks
     hunks = list(torch.split(sequences_in_row, split_sizes))
     # append EOS token to sequences
-    add_eos_tokens(hunks, eos_index)
+    add_eos_tokens(hunks, eos_index, device)
     # add padding
     hunks = pad_sequence(hunks, batch_first=True,
                          padding_value=pad_index)
@@ -32,7 +32,7 @@ def split_sequences_on_hunks(sequences: Tensor, new_hunk_index: int,
     return hunks, lengths
 
 
-def add_eos_tokens(hunks: List[Tensor], eos_index: int):
-    eos_token = torch.tensor([eos_index])
+def add_eos_tokens(hunks: List[Tensor], eos_index: int, device):
+    eos_token = torch.tensor([eos_index]).to(device)
     for i, sequence_hunk in enumerate(hunks):
         hunks[i] = torch.cat((sequence_hunk, eos_token))
